Fortune Commit: ===  ALL USERS PLEASE NOTE  ========================
Compiler optimizations have been made to macro expand LET into a WITHOUT-
INTERRUPTS special form so that it can PUSH things into a stack in the
LET-OPTIMIZATION area, SETQ the variables and then POP them back when it's
done.  Don't worry about this unless you use multiprocessing.
Note that LET *could* have been defined by:

	(LET ((LET '`(LET ((LET ',LET))
			,LET)))
	`(LET ((LET ',LET))
		,LET))

This is believed to speed up execution by as much as a factor of 1.01 or
3.50 depending on whether you believe our friendly marketing representatives.
This code was written by a new programmer here (we snatched him away from
Itty Bitti Machines where he was writing COUGHBOL code) so to give him
confidence we trusted his vows of "it works pretty well" and installed it.